Concluidas as funcionalidades de criar, remover, editar, buscar, buscar todas, buscar por tipo de pagamento, buscar to, buscar por cliente, buscar por data, fechar e reabrir referente a c, remover e buscar pedidos

+++ src/main/java/br/com/totemAutoatendimento/infraestrutura/persistencia/springdata/mysql/comanda/ComandaEntity.java
@@ -59,7 +59,7 @@ public class ComandaEntity {
 
 	private BigDecimal valor;
 
-	private Integer desconto;
+	private Float desconto;
 
 	public ComandaEntity(Comanda comanda) {
 		this.id = comanda.getId();
